extracting `svm_storage2` crate layout code to new `svm-layout` crate.
