+++ crates/svm-layout/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "svm-storage2"
+name = "svm-layout"
 version = "0.0.0"
 authors = ["Yaron Wittenstein <yaron.wittenstein@gmail.com>"]
 license = "MIT"
@@ -10,11 +10,4 @@ homepage = "https://github.com/spacemeshos/svm"
 description = "Spacemesh Virtual Machine"
 publish = false
 
-[dependencies.svm-kv]
-path = "../svm-kv"
-
-[dependencies.svm-common]
-path = "../svm-common"
-
-[dependencies.svm-layout]
-path = "../svm-layout"
+[dependencies]
