--- crates/svm-storage2/src/layout/mod.rs
@@ -1,5 +1,0 @@
-mod builder;
-mod layout;
-
-pub use builder::DataLayoutBuilder;
-pub use layout::{DataLayout, VarId};
